remove structure map integration => retired
