+++ samples/AspNetSample/AppTenantResolver.cs
