--- samples/AspNetSample/project.json
@@ -1,57 +1,0 @@
-{
-	"version": "1.0.0-*",
-	"dependencies": {
-		"Microsoft.AspNetCore.Server.Kestrel": "1.0.0",
-		"Microsoft.Extensions.Logging.Console": "1.0.0",
-		"Microsoft.Extensions.Logging.Debug": "1.0.0",
-		"Microsoft.AspNetCore.Server.IISIntegration": "1.0.0",
-		"SaasKit.Multitenancy": { "target": "project" },
-		"Microsoft.VisualStudio.Web.CodeGeneration.Tools": {
-			"version": "1.0.0-preview2-final",
-			"type": "build"
-		}
-	},
-	"frameworks": {
-		"net46": { },
-		"netcoreapp1.0": {
-			"imports": [
-				"dotnet5.6",
-				"portable-net45+win8"
-			],
-			"dependencies": {
-				"Microsoft.NETCore.App": {
-					"version": "1.0.0",
-					"type": "platform"
-				}
-			}
-		}
-	},
-	"tools": {
-		"BundlerMinifier.Core": "2.0.238",
-		"Microsoft.AspNetCore.Server.IISIntegration.Tools": "1.0.0-preview2-final"
-	},
-    "buildOptions": {
-        "emitEntryPoint": true,
-        "preserveCompilationContext": true
-    },
-
-    "runtimeOptions": {
-        "configProperties": {
-            "System.GC.Server": true
-        }
-    },
-
-	"publishOptions": {
-        "include": [
-            "wwwroot",
-            "Views",
-            "Areas/**/Views",
-            "appsettings.json",
-            "web.config"
-        ]
-	},
-	"scripts": {
-		"prepublish": [ "bower install", "dotnet bundle" ],
-		"postpublish": "dotnet publish-iis --publish-folder %publish:OutputPath% --framework %publish:FullTargetFramework%"
-	}
-}
