+++ samples/AspNetStructureMapSample/AspNetStructureMapSample.csproj
@@ -1,0 +1,11 @@
+﻿<Project Sdk="Microsoft.NET.Sdk.Web">
+
+	<PropertyGroup>
+		<TargetFramework>net6.0</TargetFramework>
+		<Platforms>AnyCPU;x64</Platforms>
+	</PropertyGroup>
+
+	<ItemGroup>
+	  <ProjectReference Include="..\..\src\SaasKit.Multitenancy.StructureMap\SaasKit.Multitenancy.StructureMap.csproj" />
+	</ItemGroup>
+</Project>
